+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/builtins/modules/PolyglotModuleBuiltins.java
@@ -231,21 +231,38 @@ public final class PolyglotModuleBuiltins extends PythonBuiltins {
         }
     }
 
-    @Builtin(name = "export_value", minNumOfPositionalArgs = 1, parameterNames = {"value", "name"})
+    @Builtin(name = "export_value", minNumOfPositionalArgs = 1, parameterNames = {"name", "value"})
     @GenerateNodeFactory
     public abstract static class ExportSymbolNode extends PythonBuiltinNode {
         @Child private GetAttributeNode getNameAttributeNode;
         @Child private CastToStringNode castToStringNode;
 
-        @Specialization
+        @Specialization(guards = "!isString(value)")
         @TruffleBoundary
-        public Object exportSymbol(Object value, String name) {
+        public Object exportSymbolKeyValue(String name, Object value) {
             Env env = getContext().getEnv();
             if (!env.isPolyglotAccessAllowed()) {
                 throw raise(PythonErrorType.NotImplementedError, "polyglot access is not allowed");
             }
             env.exportSymbol(name, value);
             return value;
+        }
+
+        @Specialization(guards = "!isString(value)")
+        @TruffleBoundary
+        public Object exportSymbolValueKey(Object value, String name) {
+            PythonLanguage.getLogger().warning("[deprecation] polyglot.export_value(value, name) is deprecated " +
+                            "and will be removed. Please swap the arguments.");
+            return exportSymbolKeyValue(name, value);
+        }
+
+        @Specialization(guards = "isString(arg1)")
+        @TruffleBoundary
+        public Object exportSymbolAmbiguous(Object arg1, String arg2) {
+            PythonLanguage.getLogger().warning("[deprecation] polyglot.export_value(str, str) is ambiguous. In the future, this will " +
+                            "default to using the first argument as the name and the second as value, but now it " +
+                            "uses the first argument as value and the second as the name.");
+            return exportSymbolValueKey(arg1, arg2);
         }
 
         @Specialization